refactor: seperate Input Component
1. whenever user type input, dashboard re-rendered.
2. So I seperated Componenet
