+++ src/components/UI/index.ts
@@ -2,10 +2,12 @@ import DashBoard from './Dashboard';
 import Input from './Input';
 import Output from './Output';
 import Loading from './Loading';
+import InputBox from './InputBox';
 
 export {
   DashBoard,
   Input,
   Output,
-  Loading
+  Loading,
+  InputBox
 };
